Initial working version of version 2 audio units support prior to cleanup.
This contains commented out failed attempts, false starts and duplications of existing functionality conforming to both the V1 (Component Manager) and V2 (Audio Component) interfaces.
Passes AU validation and appears to work correctly in both 64-bit and 32-bit builds. However for the 32-bit build the GUI has been forced to be Cocoa rather than the default of Carbon and does have visual glitches. Having a Cocoa GUI for the 32-bit architecture as well as the 64-bit architecture means less stuff to have to port over from the Component Manager. (It appears that Carbon AU UIs need to be supplied via the Component Manager. Rather than spend time investigating how to work around that, we may as well take advantage of the fact that macOS has been a mainly Cocoa world for some time now.)