Unwind the complex intertwining of template classes, in-memory constructors and other indirections copied in from Apple's AU SDK.
All that is really needed is a factory class that creates for each plugin instance a wrapper struct containing the `AudioComponentPlugInInterface` struct expected of an AU v2 factory plus extra fields that are easily accessed when that `AudioComponentPlugInInterface` is passed back in via the callbacks defined via the selector lookup. This provides access to the pointer to the actual C++ implementation object.