Fix component name handling
Earlier commits removed the `getComponentName()` member function because the NUsofting branch no longer needed the component name apart from Symbiosis using it for the global trace identifier. However, the VST preset conversion code needs the component name, so this now needs to be fed in to the `SymbiosisComponent` C++ object via its constructor. Took the opportunity to modernise the AU v1 and v2 helper functions that extract the component name from the component instances. In the process, found a bug in the 32-bit AU v2 handling of component names, now fixed.